requirements.txt
numpy>=1.23.0
opencv-python>=4.8.0
Pillow>=10.0.0
